Move `_resolve_one_strong_one_weak_types` and `_resolve_one_strong_two_weak_types` to `_type_utils`
